pyproject.toml
[project]
name = "mlbb-bot-server"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot[webhooks]>=21.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
]

[tool.setuptools.packages.find]
include = ["bot*", "agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["bot", "agent", "config"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 100
show_missing = true
